modify item received module

+++ app/Http/Controllers/Admin/ItemReceivedController.php
@@ -57,7 +57,10 @@ class ItemReceivedController extends Controller
         //Multiply Quantity and Unite Rate to get total amount of the items received
         $request['amount'] = multiply_two_numbers($request->quantity,$request->unit_rate);
 
-        $itemReceived->create($request->all());
+        $itemReceived = $itemReceived->create($request->all());
+
+        //Increment the Item Quantity
+        $this->item_quantity($itemReceived->item_id, $itemReceived->quantity);
 
         return back()->with('success','Item has been received');
     }
@@ -88,6 +91,12 @@ class ItemReceivedController extends Controller
 
         $itemReceived->update($request->all());
 
+        //Check if Item Quantity Updated
+        if($request->quantity != $request->old_quantity)
+        {
+            $this->update_item_quantity($request->item_id,$request->quantity,$request->old_quantity);
+        }
+
         return redirect()->route('itemReceived.index')->with('success','Received Item has been Updated');
     }
 
@@ -97,10 +106,70 @@ class ItemReceivedController extends Controller
      */
     public function destroy(ItemReceived $itemReceived)
     {
-        $itemReceived->delete();
+        if($this->check_before_delete_item_received ($itemReceived->item_id,$itemReceived->quantity) == false) {
+            $itemReceived->delete();
+            return redirect()->route('itemReceived.index')->with('success','Received Item has been deleted');
+        }
+        else {
+            return redirect()->route('itemReceived.index')->with('error','Received Item can not be deleted');
+        }
 
-        return redirect()->route('itemReceived.index')->with('success','Received Item has been deleted');
     }
+
+    /*
+     * Increment Item Quantity Number
+     */
+    public function item_quantity($item_id, $quantity)
+    {
+        $item = $this->get_item_id($item_id);   //Initialize the class
+
+        $item->increment('quantity',$quantity);
+
+        return null; //return nothing
+    }
+
+    /*
+     * Update the Item Quantity
+     */
+    public function update_item_quantity($item_id,$quantity,$old_quantity)
+    {
+        $item = $this->get_item_id($item_id);
+
+        $item->decrement('quantity',$old_quantity);
+
+        $item->increment('quantity',$quantity);
+
+        return null; //return nothing
+    }
+
+    /*
+     * Delete the Item Quantity
+     */
+    public function check_before_delete_item_received ($item_id, $quantity)
+    {
+        $item = $this->get_item_id($item_id);
+
+        if($item->quantity < $quantity )
+        {
+            return true;
+        }
+        else {
+            $item->decrement('quantity',$quantity);
+
+            return false;
+        }
+    }
+
+    /*
+     * Get Item ID
+     */
+    public function get_item_id ($item)
+    {
+        $item = Item::where('id',$item)->first();
+
+        return $item;
+    }
+
 
 
 }
